centervalue {RSAGA}	R Documentation

Pick Center Value from Matrix

Description

Pick the value in the center of a square matrix. Auxiliary function to be used by functions called by focal.function.

Usage

centervalue(x)

Arguments

x
a square matrix

Details

See for example the code of resid.median.

See Also

focal.function, resid.median

Examples

( m <- matrix( round(runif(9,1,10)), ncol=3 ) )
centervalue(m)

[Package RSAGA version 0.94-5 Index]