rsaga.add.grid.values.to.points {RSAGA}	R Documentation

Add Grid Values to Point Shapefile

Description

Pick values from SAGA grids and attach them as a new variables to a point shapefile.

Usage

rsaga.add.grid.values.to.points(in.shapefile, in.grids, out.shapefile,
  method = c("nearest.neighbour", "bilinear", "idw", "bicubic.spline",
  "b.spline"), ...)

Arguments

in.shapefile
Input point shapefile (default extension: .shp).

in.grids
Input: character vector with names of (one or more) SAGA GIS grid files to be converted into a point shapefile.

out.shapefile
Output point shapefile (default extension: .shp).

method
interpolation method to be used; choices: nearest neighbour interpolation (default), bilinear interpolation, inverse distance weighting, bicubic spline interpolation, B-splines.

...
Optional arguments to be passed to rsaga.geoprocessor, including the env RSAGA geoprocessing environment.

Details

Retrieves information from the selected grids at the positions of the points of the selected points layer and adds it to the resulting layer.

Note

This function uses module Add Grid Values to Points in SAGA GIS library shapes_grid.

Author(s)

Alexander Brenning (R interface), Olaf Conrad (SAGA modules)

See Also

pick.from.points, pick.from.ascii.grid, pick.from.saga.grid, rsaga.grid.to.points

[Package RSAGA version 0.94-5 Index]