rsaga.close.gaps {RSAGA}	R Documentation

SAGA Modules Close Gaps and Close One Cell Gaps

Description

Close (Interpolate) Gaps

Usage

rsaga.close.gaps(in.dem, out.dem, threshold = 0.1, ...)

rsaga.close.one.cell.gaps(in.dem, out.dem, ...)

Arguments

in.dem
input: digital elevation model (DEM) as SAGA grid file (default file extension: .sgrd)

out.dem
output: DEM grid file without no-data values (gaps). Existing files will be overwritten!

threshold
tension threshold for adjusting the interpolator (default: 0.1)

...
optional arguments to be passed to rsaga.geoprocessor, including the env RSAGA geoprocessing environment

Details

rsaga.close.one.cell.gaps only fill gaps whose neighbor grid cells have non-missing data.

In rsaga.close.gaps, larger tension thresholds can be used to reduce overshoots and undershoots in the surfaces used to fill (interpolate) the gaps.

Value

The type of object returned depends on the intern argument passed to the rsaga.geoprocessor. For intern=FALSE it is a numerical error code (0: success), or otherwise (default) a character vector with the module's console output.

Note

This function uses modules 7 (rsaga.close.gaps and 6 rsaga.close.one.cell.gaps from the SAGA library grid_tools.

SAGA GIS 2.0.5+ has a new additional module Close Gaps with Spline, which can be accessed using rsaga.geoprocessor (currently no R wrapper available). See rsaga.get.usage("grid_tools","Close Gaps with Spline") or in version 2.1.0+ call rsaga.html.help("grid_tools","Close Gaps with Spline").

Author(s)

Alexander Brenning (R interface), Olaf Conrad (SAGA module)

See Also

rsaga.geoprocessor, rsaga.env

Examples

## Not run: 
# using SAGA grids:
rsaga.close.gaps("rawdem.sgrd","dem.sgrd")
# using ASCII grids:
rsaga.esri.wrapper(rsaga.close.gaps,in.dem="rawdem",out.dem="dem")

## End(Not run)

[Package RSAGA version 0.94-5 Index]