rsaga.contour {RSAGA}	R Documentation

Contour Lines from a Grid

Description

Creates a contour lines shapefile from a grid file in SAGA grid format.

Usage

rsaga.contour(in.grid, out.shapefile, zstep, zmin, zmax, vertex = "xy",
  env = rsaga.env(), ...)

Arguments

in.grid
input: digital elevation model (DEM) as SAGA grid file (default file extension: .sgrd)

out.shapefile
output: contour line shapefile. Existing files will be overwritten!

zstep,zmin,zmax
lower limit, upper limit, and equidistance of contour lines

vertex
optional parameter: vertex type for resulting contours. Default "xy" (or 0). Only available with SAGA GIS 2.1.3+.

[0] "xy"

[1] "xyz"

env
A SAGA geoprocessing environment, see rsaga.env

...
arguments to be passed to rsaga.geoprocessor

Value

The type of object returned depends on the intern argument passed to the rsaga.geoprocessor. For intern=FALSE it is a numerical error code (0: success), or otherwise (the default) a character vector with the module's console output.

Author(s)

Alexander Brenning (R interface), Olaf Conrad (SAGA module)

See Also

rsaga.geoprocessor

[Package RSAGA version 0.94-5 Index]