rsaga.copy.sgrd {RSAGA}	R Documentation

Create a copy of a SAGA grid file

Description

Creates a copy of a SAGA grid file, optionally overwriting the target file if it already exists. Intended mainly for internal use by RSAGA functions, currently in particular rsaga.inverse.distance.

Usage

rsaga.copy.sgrd(in.grid, out.grid, overwrite = TRUE, env = rsaga.env())

Arguments

in.grid
name of a SAGA GIS grid file; file extension can be omitted

out.grid
name of a SAGA GIS grid file; file extension can be omitted

overwrite
logical; if TRUE (the default), overwrite out.grid if it already exists; if FALSE and the out.grid already exists, copying will be skipped without causing an error.

env
a SAGA geoprocessing environment as created by rsaga.env

Note

SAGA grid files consist of three (or more) individual files with file extensions .mgrd, .sgrd and .sdat. The files with these three file extensions are copied, any additional files (e.g. a history file) are ignored.

[Package RSAGA version 0.94-5 Index]