rsaga.filter.simple {RSAGA}	R Documentation

Simple Filters

Description

Apply a smoothing, sharpening or edge filter to a SAGA grid.

Usage

rsaga.filter.simple(in.grid, out.grid, mode = "circle", method = c("smooth",
  "sharpen", "edge"), radius, ...)

Arguments

in.grid
input: SAGA grid file (default file extension: .sgrd)

out.grid
output: SAGA grid file

mode
character or numeric: shape of moving window, either "square" (=0) or "circle" (=1, default)

method
character or numeric: "smooth" (=0), "sharpen" (=1), or "edge" (=2)

radius
positive integer: radius of moving window

...
optional arguments to be passed to rsaga.geoprocessor, including the env RSAGA geoprocessing environment

Value

The type of object returned depends on the intern argument passed to the rsaga.geoprocessor. For intern=FALSE it is a numerical error code (0: success), or otherwise (the default) a character vector with the module's console output.

Author(s)

Alexander Brenning (R interface), Olaf Conrad (SAGA module)

See Also

rsaga.filter.gauss

Examples

## Not run: rsaga.filter.simple("dem","dem-smooth",radius=4)

[Package RSAGA version 0.94-5 Index]