rsaga.get.usage {RSAGA}	R Documentation

Usage of SAGA command line modules

Description

rsaga.get.usage provides information on the usage of and arguments required by SAGA command line modules.

Usage

rsaga.get.usage(lib, module, env = rsaga.env(), show = TRUE)

Arguments

lib
name of the SAGA library

module
name or numeric identifier of SAGA module in library lib

env
a SAGA geoprocessing environment as created by rsaga.env

show
logical (default: TRUE); display usage in the R console?

Details

This function is intended to provide information required to use the rsaga.geoprocessor and for writing your own high-level interface function for SAGA modules. R–SAGA interfaces already exist for some SAGA modules, e.g. rsaga.hillshade, rsaga.local.morphometry, but there are many more.

Value

The character vector with usage information is invisibly returned.

See Also

rsaga.html.help, rsaga.geoprocessor, rsaga.env, rsaga.get.modules

Examples

## Not run: 
rsaga.get.usage("io_grid",1)
rsaga.get.usage("ta_preprocessor",2)
rsaga.get.usage("ta_morphometry",0)
# in SAGA GIS 2.1.0+, compare:
rsaga.html.help("io_grid",1)
# etc.

## End(Not run)

[Package RSAGA version 0.94-5 Index]