rsaga.hillshade {RSAGA}	R Documentation

Analytical hillshading Analytical hillshading calculation.

Description

Analytical hillshading Analytical hillshading calculation.

Usage

rsaga.hillshade(in.dem, out.grid, method = "standard", azimuth = 315,
  declination = 45, exaggeration = 4, ...)

Arguments

in.dem
Input digital elevation model (DEM) as SAGA grid file (default extension: .sgrd).

out.grid
Output hillshading grid (SAGA grid file). Existing files will be overwritten!

method
Available choices (character or numeric): "standard" (or 0 - default), "max90deg.standard" (1), "combined.shading" (2), "ray.tracing" (3). See Details.

azimuth
Direction of the light source, measured in degree clockwise from the north direction; default 315, i.e. northwest.

declination
Declination of the light source, measured in degree above the horizon (default 45).

exaggeration
Vertical exaggeration of elevation (default: 4). The terrain exaggeration factor allows to increase the shading contrasts in flat areas.

...
Optional arguments to be passed to rsaga.geoprocessor, including the env RSAGA geoprocessing environment.

Details

The Analytical Hillshading algorithm is based on the angle between the surface and the incoming light beams, measured in radians.

Value

The type of object returned depends on the intern argument passed to the rsaga.geoprocessor. For intern=FALSE it is a numerical error code (0: success), or otherwise (default) a character vector with the module's console output.

Note

While the default azimuth of 315 degree (northwest) is not physically meaningful on the northern hemisphere, a northwesterly light source is required to properly depict relief in hillshading images. Physically correct southerly light sources results a hillshade that would be considered by most people as inverted: hills look like depressions, mountain chains like troughs.

Author(s)

Alexander Brenning (R interface), Olaf Conrad (SAGA module)

See Also

rsaga.solar.radiation, rsaga.insolation

Examples

## Not run: rsaga.hillshade("dem.sgrd","hillshade")

[Package RSAGA version 0.94-5 Index]