rsaga.import.gdal {RSAGA}	R Documentation

Import Grid Files to SAGA grid format using GDAL

Description

These functions provide simple interfaces for reading and writing grids from/to ASCII grids and Rd files. Grids are stored in matrices, their headers in lists.

Usage

rsaga.import.gdal(in.grid, out.grid, env = rsaga.env(), ...)

Arguments

in.grid
file name of a grid in a format supported by GDAL

out.grid
output SAGA grid file name; defaults to in.grid with the file extension being removed; file extension should not be specified, it defaults to .sgrd

env
RSAGA geoprocessing environment created by rsaga.env

...
additional arguments to be passed to rsaga.geoprocessor

Details

The GDAL Raster Import module of SAGA imports grid data from various file formats using the Geospatial Data Abstraction Library (GDAL) by Frank Warmerdam. GDAL Versions are specific to SAGA versions:

SAGA 2.0.7 - 2.0.8: GDAL v.1.8.0

SAGA 2.1.0 - 2.1.1: GDAL v.1.10.0

SAGA 2.1.2 - 2.2.0: GDAL v.1.11.0

SAGA 2.2.1 - 2.2.3: GDAL v.2.1.0 dev

More information is available at http://www.gdal.org/.

If in.grid has more than one band (e.g. RGB GEOTIFF), then output grids with file names of the form in.grid_01.sgrd, in.grid_02.sgrd etc. are written, one for each band.

The following raster formats are currently supported. Last updated for SAGA GIS 2.2.3; for a list for a specific SAGA GIS version call rsaga.html.help("io_gdal","GDAL: Import Raster", env = rsaga.env(path="SAGA_Version_to_Test"))

BAG - Bathymetry Attributed Grid

ECW - ERDAS Compressed Wavelets (SDK 3.x)

JP2ECW - ERDAS JPEG2000 (SDK 3.x)

FITS - Flexible Image Transport System

GMT - GMT NetCDF Grid Format

HDF4 - Hierarchical Data Format Release 4

HDF4Image - HDF4 Dataset

HDF5 - Hierarchical Data Format Release 5

HDF5Image - HDF5 Dataset

KEA - KEA Image Format (.kea)

MG4Lidar - MrSID Generation 4 / Lidar (.sid)

MrSID - Multi-resolution Seamless Image Database (MrSID)

netCDF - Network Common Data Format

PostgreSQL - PostgreSQL/PostGIS

VRT - Virtual Raster

GTiff - GeoTIFF

NITF - National Imagery Transmission Format

RPFTOC - Raster Product Format TOC format

ECRGTOC - ECRG TOC format

HFA - Erdas Imagine Images (.img)

SAR_CEOS - CEOS SAR Image

CEOS - CEOS Image

JAXAPALSAR - JAXA PALSAR Product Reader (Level 1.1/1.5)

GFF - Ground-based SAR Applications Testbed File Format (.gff)

ELAS - ELAS

AIG - Arc/Info Binary Grid

AAIGrid - Arc/Info ASCII Grid

GRASSASCIIGrid - GRASS ASCII Grid

SDTS - SDTS Raster

DTED - DTED Elevation Raster

PNG - Portable Network Graphics

JPEG - JPEG JFIF

MEM - In Memory Raster

JDEM - Japanese DEM (.mem)

GIF - Graphics Interchange Format (.gif)

BIGGIF - Graphics Interchange Format (.gif)

ESAT - Envisat Image Format

BSB - Maptech BSB Nautical Charts

XPM - X11 PixMap Format

BMP - MS Windows Device Independent Bitmap

DIMAP - SPOT DIMAP

AirSAR - AirSAR Polarimetric Image

RS2 - RadarSat 2 XML Product

SAFE - Sentinel SAFE Product

PCIDSK - PCIDSK Database File

PCRaster - PCRaster Raster File

ILWIS - ILWIS Raster Map

SGI - SGI Image File Format 1.0

SRTMHGT - SRTMHGT File Format

Leveller - Leveller heightfield

Terragen - Terragen heightfield

ISIS3 - USGS Astrogeology ISIS cube (Version 3)

ISIS2 - USGS Astrogeology ISIS cube (Version 2)

PDS - NASA Planetary Data System

VICAR - MIPL VICAR file

TIL - EarthWatch .TIL

ERS - ERMapper .ers Labelled

JP2OpenJPEG - JPEG-2000 driver based on OpenJPEG library

L1B - NOAA Polar Orbiter Level 1b Data Set

FIT - FIT Image

GRIB - GRIdded Binary (.grb)

RMF - Raster Matrix Format

WCS - OGC Web Coverage Service

WMS - OGC Web Map Service

MSGN - EUMETSAT Archive native (.nat)

RST - Idrisi Raster A.1

INGR - Intergraph Raster

GSAG - Golden Software ASCII Grid (.grd)

GSBG - Golden Software Binary Grid (.grd)

GS7BG - Golden Software 7 Binary Grid (.grd)

COSAR - COSAR Annotated Binary Matrix (TerraSAR-X)

TSX - TerraSAR-X Product

COASP - DRDC COASP SAR Processor Raster

R - R Object Data Store

MAP - OziExplorer .MAP

PNM - Portable Pixmap Format (netpbm)

DOQ1 - USGS DOQ (Old Style)

DOQ2 - USGS DOQ (New Style)

ENVI - ENVI .hdr Labelled

EHdr - ESRI .hdr Labelled

GenBin - Generic Binary (.hdr Labelled)

PAux - PCI .aux Labelled

MFF - Vexcel MFF Raster

MFF2 - Vexcel MFF2 (HKV) Raster

FujiBAS - Fuji BAS Scanner Image

GSC - GSC Geogrid

FAST - EOSAT FAST Format

BT - VTP .bt (Binary Terrain) 1.3 Format

LAN - Erdas .LAN/.GIS

CPG - Convair PolGASP

IDA - Image Data and Analysis

NDF - NLAPS Data Format

EIR - Erdas Imagine Raw

DIPEx - DIPEx

LCP - FARSITE v.4 Landscape File (.lcp)

GTX - NOAA Vertical Datum .GTX

LOSLAS - NADCON .los/.las Datum Grid Shift

NTv2 - NTv2 Datum Grid Shift

CTable2 - CTable2 Datum Grid Shift

ACE2 - ACE2

SNODAS - Snow Data Assimilation System

KRO - KOLOR Raw

ROI_PAC - ROI_PAC raster

ISCE - ISCE raster

ARG - Azavea Raster Grid format

RIK - Swedish Grid RIK (.rik)

USGSDEM - USGS Optional ASCII DEM (and CDED)

GXF - GeoSoft Grid Exchange Format

NWT_GRD - Northwood Numeric Grid Format .grd/.tab

NWT_GRC - Northwood Classified Grid Format .grc/.tab

ADRG - ARC Digitized Raster Graphics

SRP - Standard Raster Product (ASRP/USRP)

BLX - Magellan topo (.blx)

Rasterlite - Rasterlite

PostGISRaster - PostGIS Raster driver

SAGA - SAGA GIS Binary Grid (.sdat)

KMLSUPEROVERLAY - Kml Super Overlay

XYZ - ASCII Gridded XYZ

HF2 - HF2/HFZ heightfield raster

PDF - Geospatial PDF

OZI - OziExplorer Image File

CTG - USGS LULC Composite Theme Grid

E00GRID - Arc/Info Export E00 GRID

ZMap - ZMap Plus Grid

NGSGEOID - NOAA NGS Geoid Height Grids

MBTiles - MBTiles

IRIS - IRIS data (.PPI, .CAPPi etc)

PLMOSAIC - Planet Labs Mosaic

CALS - CALS (Type 1)

WMTS - OGC Web Map Tile Service

ESRI Shapefile - ESRI Shapefile

MapInfo File - MapInfo File

UK .NTF - UK .NTF

OGD_SDTS - SDTS

S57 - IHO S-57 (ENC)

DGN - Microstation DGN

OGR_VRT - VRT - Virtual Datasource

REC EPIInfo .REC

Memory - Memory

BNA - Atlas BNA

CSV - Comma Separated Value (.csv)

NAS - NAS - ALKIS

GML - Geography Markup Language

GPX - GPX

LIBKML - Keyhole Markup Language (LIBKML)

KML - Keyhole Markup Language (KML)

GeoJSON - GeoJSON

Interlis 1 - Interlis 1

Interlis 2 - Interlis 2

OGR_GMT - GMT ASCII Vectors (.gmt)

GPKG - GeoPackage

SQLite - SQLite / Spatialite

ODBC - ODBC

WAsP - WAsP .map format

PGeo - ESRI Personal GeoDatabase

MSSQLSpatial - Microsoft SQL Server Spatial Database

MySQL - MySQL

OpenFileGDB - ESRI FileGDB

XPlane - X-Plane/Flightgear aeronautical data

DXF - AutoCAD DXF

Geoconcept - Geoconcept

GeoRSS - GeoRSS

GPSTrackMaker - GPSTrackMaker

VFK - Czech Cadastral Exchange Data Format

PGDUMP - PostgreSQL SQL dump

OSM - OpenStreetMap XML and PDF

GPSBabel - GPSBabel

SUA - Tim Newport-Peace's Special Use Airspace Format

OpenAir - OpenAir

OGR_PDS - Planetary Data Systems TABLE

WFS - OGC WFS (Web Feature Service)

HTF - Hydrographic Transfer Vector

AeronavFAA - Aeronav FAA

Geomedia - Geomedia .mdb

EDIGEO - French EDIGEO exchange format

GFT - Google Fusion Tables

GME - Google Maps Engine

SVG - Scalable Vector Graphics

CouchDB - CouchDB / GeoCouch

Cloudant - Cloudant / CouchDB

Idrisi - Idrisi Vector (.vct)

ARCGEN - Arc/Info Generate

SEGUKOOA - SEG-P1 / UKOOA P1/90

SEG-Y - SEG-Y

ODS - Open Document/ LibreOffice / OpenOffice Spreadsheet

XLSX - MS Office Open XML spreadsheet

ElasticSearch - Elastic Search

Walk - Walk

CartoDB - CartoDB

SXF - Storage and eXchange Format

Selafin - Selafin

JML - OpenJUMP JML

PLSCENES - Planet Labs Scenes API

CSW - OGC CSW (Catalog Search for the Web)

IDF - INTREST Data Format

TIGER - U.S. Census TIGER/Line

AVCBin - Arc/Info Binary Coverage

AVCE00 - Arc/Info E00 (ASCII) Coverage

HTTP - HTTP Fetching Wrapper

Author(s)

Alexander Brenning (R interface), Olaf Conrad / Andre Ringeler (SAGA module), Frank Warmerdam (GDAL)

References

GDAL website: http://www.gdal.org/

See Also

read.ascii.grid, rsaga.esri.to.sgrd, read.sgrd, read.Rd.grid

[Package RSAGA version 0.94-5 Index]