rsaga.lib.prefix {RSAGA}	R Documentation

Determine prefix for SAGA GIS library names

Description

Internal function that determines the possible prefix for SAGA GIS library names - relevant for non-Windows SAGA GIS pre-2.1.0.

Usage

rsaga.lib.prefix(env)

Arguments

env
list, setting up a SAGA geoprocessing environment as created by rsaga.env.

Details

Some non-Windows versions of saga_cmd require library names with a "lib" prefix, e.g. libio_grid instead of io_grid. This function, which is called by rsaga.env tries to guess this behaviour based on the operating system and SAGA GIS version.

Value

A character string, either "" or "lib".

See Also

rsaga.env

Examples

## Not run: 
env = rsaga.env()
# obtained by a call to rsaga.lib.prefix:
env$lib.prefix

# more explicitly:
rsaga.lib.prefix(env=env)

## End(Not run)

[Package RSAGA version 0.94-5 Index]