rsaga.sink.removal {RSAGA}	R Documentation

Sink Removal Remove sinks from a digital elevation model by deepening drainage routes or filling sinks.

Description

Sink Removal Remove sinks from a digital elevation model by deepening drainage routes or filling sinks.

Usage

rsaga.sink.removal(in.dem, in.sinkroute, out.dem, method = "fill", ...)

Arguments

in.dem
input: digital elevation model (DEM) as SAGA grid file (default file extension: .sgrd)

in.sinkroute
optional input: sink route grid file

out.dem
output: modified DEM

method
character string or numeric value specifying the algorithm (partial string matching will be applied): "deepen drainage route" (or 0): reduce the elevation of pixels in order to achieve drainage out of the former sinks "fill sinks" (or 1): fill sinks until none are left

...
optional arguments to be passed to rsaga.geoprocessor, including the env RSAGA geoprocessing environment

Value

The type of object returned depends on the intern argument passed to the rsaga.geoprocessor. For intern=FALSE it is a numerical error code (0: success), or otherwise (default) a character vector with the module's console output.

Note

This function uses module 1 from SAGA library ta_preprocessor.

Author(s)

Alexander Brenning (R interface), Olaf Conrad (SAGA module)

See Also

rsaga.sink.route, rsaga.fill.sinks

Examples

## Not run: rsaga.sink.route("dem","sinkroute")
rsaga.sink.removal("dem","sinkroute","dem-preproc",method="deepen")
## End(Not run)

[Package RSAGA version 0.94-5 Index]