rsaga.sink.route {RSAGA}	R Documentation

Sink Drainage Route Detection

Description

Sink drainage route detection.

Usage

rsaga.sink.route(in.dem, out.sinkroute, threshold, thrsheight = 100, ...)

Arguments

in.dem
input: digital elevation model (DEM) as SAGA grid file (default file extension: .sgrd)

out.sinkroute
output: sink route grid file: non-sinks obtain a value of 0, sinks are assigned an integer between 0 and 8 indicating the direction to which flow from this sink should be routed

threshold
logical: use a threshold value?

thrsheight
numeric: threshold value (default: 100)

...
optional arguments to be passed to rsaga.geoprocessor, including the env RSAGA geoprocessing environment

Value

The type of object returned depends on the intern argument passed to the rsaga.geoprocessor. For intern=FALSE it is a numerical error code (0: success), or otherwise (default) a character vector with the module's console output.

Note

I assume that flow directions are coded as 0 = north, 1 = northeast, 2 = east, ..., 7 = northwest, as in rsaga.fill.sinks.

Author(s)

Alexander Brenning (R interface), Olaf Conrad (SAGA module)

See Also

rsaga.sink.removal

Examples

## Not run: rsaga.sink.route("dem","sinkroute")
rsaga.sink.removal("dem","sinkroute","dem-preproc",method="deepen")
## End(Not run)

[Package RSAGA version 0.94-5 Index]