set.file.extension {RSAGA}	R Documentation

Determine or modify file name extensions

Description

Function get.file.extension determines the file extension, set.file.extension changes it, and default.file.extension changes it only if it is not already specified.

Usage

set.file.extension(filename, extension, fsep = .Platform$file.sep)

get.file.extension(filename, fsep = .Platform$file.sep)

default.file.extension(filename, extension, force = FALSE)

Arguments

filename
character vector: file name(s), possibly including paths and extensions; a file name ending with a "." is interpreted as having extension "", while a file name that doesn't contain a "." is interpreted has having no extension.

extension
character string: file extension, without the dot

fsep
character: separator between paths

force
logical argument to default.file.extension: force the file extension to be extension (same result as set.file.extension), or only set it to extension if it has not been specified?

Value

character vector of same length as filename

Examples

fnm = c("C:/TEMP.DIR/temp","C:/TEMP.DIR/tmp.txt","tempfile.")
get.file.extension(fnm)
set.file.extension(fnm,extension=".TMP")
default.file.extension(fnm,extension=".TMP")

[Package RSAGA version 0.94-5 Index]